feat(server): migrate Express → Bun.serve + Hono (#316)
Cuts the server over entirely. The spike branch proved the Hono build boots and
serves the full app in a browser; this commit makes it the ONLY HTTP layer by
deleting the Express code and dropping its dependencies.

What changed:
- index.ts is now the Bun.serve + Hono entrypoint (was Express). The AppContext
  type export is preserved verbatim, so every `import type { AppContext } from
  #/index` (services, auth, hono/*) keeps resolving with zero churn. The old
  Express `Server` class was imported by nothing; the new Hono Server replaces it.
- Deleted: src/index-hono.ts (folded into index.ts), src/routes.ts, src/routes/
  (6 files), src/controllers/ (5 files).
- src/auth/session.ts: dropped the `import cookie-session` side-effect import
  and the `declare global { namespace Express.Request.session }` augmentation
  (both Express-era residue). AppSessionData/AccountEntry types + helpers kept.
- src/auth/session-agent.ts: initializeAgentFromSession now takes a minimal
  structural `{ session?: { did?: string } | null }` instead of Express.Request,
  so the session-agent unit test stays green without the global augmentation.
- package.json: dropped express, cors, cookie-session, cookie-parser,
  express-rate-limit, express-validator (+ @types/*); removed dev:hono/start:hono
  (dev/start already run src/index.ts, now Hono).
- bunfig.toml: coverage exclusions repointed (src/routes.* removed; src/hono/**
  added; src/index.ts comment updated).
- CLAUDE.md: server architecture now two-layer (hono route handlers → services);
  auth flow + cookie-format note (Hono SHA256 single-cookie, not wire-compat with
  the old keygrip SHA1 dual-cookie — one-time re-login); controller-tests-broken
  caveat noted.

Verified:
- tsc clean.
- 88 service/lib/auth tests pass (0 fail), including session-agent.test.ts.
- Full stack boots (server :8080 + Vite :5173); landing page + login page render;
  live /handle-search?q=navy → 200 through the migrated server; curl spot-check
  across all domains (session 200, auth-gated 403, validator 400, 404).

Known/deferred (NOT in this commit):
- The 5 src/tests/*-controller.test.ts files import deleted controllers and now
  fail to import under `bun test`. Left on disk by design; rewrite is the next
  step. The 12 service/lib/auth tests are unaffected and pass.
- The real Bluesky OAuth round-trip (callback→consume→session→agent→post) is
  still unverified against a live PDS — the forged-cookie + curl tests prove the
  session mechanics, not a real login.

Dockerfile.server unchanged — its CMD already points at `bun src/index.ts`.

Refs #316.
