+++ server/package.json
@@ -9,8 +9,6 @@
   "scripts": {
     "dev": "bun --watch --no-clear-screen src/index.ts | pino-pretty",
     "start": "bun src/index.ts",
-    "dev:hono": "bun --watch --no-clear-screen src/index-hono.ts | pino-pretty",
-    "start:hono": "bun src/index-hono.ts",
     "lexgen": "lex gen-server ./src/lexicon ./lexicons/*",
     "clean": "rimraf coverage",
     "lint": "oxlint .",
@@ -38,24 +36,16 @@
     "@atproto/xrpc-server": "^0.11.6",
     "@axiomhq/pino": "^1.8.0",
     "@hono/zod-validator": "^0.9.0",
-    "@types/cookie-parser": "^1.4.10",
-    "@types/cookie-session": "^2.0.49",
     "@types/cryptr": "^4.0.3",
     "@types/web-push": "^3.6.4",
-    "cookie-parser": "^1.4.7",
-    "cookie-session": "^2.1.1",
-    "cors": "^2.8.6",
     "cryptr": "^6.4.0",
     "dotenv": "^17.4.2",
     "envalid": "^8.2.0",
-    "express": "^5.2.1",
-    "express-rate-limit": "^8.6.1",
-    "express-validator": "^7.3.2",
     "hono": "^4.12.34",
     "hono-rate-limiter": "^0.5.3",
     "iron-session": "^8.0.4",
     "kysely": "^0.29.4",
-    "multiformats": "^14.0.5",
+    "multiformats": "^14.0.4",
     "obscenity": "^0.4.6",
     "pg": "^8.22.0",
     "pino": "^10.3.1",
@@ -67,15 +57,13 @@
   "devDependencies": {
     "@atproto/lex-cli": "^0.10.6",
     "@types/bun": "1.3.14",
-    "@types/cors": "^2.8.19",
-    "@types/express": "^5.0.6",
     "@types/node": "^26.1.2",
-    "@types/pg": "^8.20.3",
+    "@types/pg": "^8.20.0",
     "@types/serve-static": "^2.2.0",
-    "oxlint": "^1.77.0",
+    "oxlint": "^1.76.0",
     "pino-pretty": "^13.1.3",
     "rimraf": "^6.1.3",
-    "tsx": "^4.23.5",
+    "tsx": "^4.22.4",
     "typescript": "^7.0.2"
   }
 }
