--- server/src/controllers/message-controller.ts
@@ -1,304 +1,0 @@
-/* v8 ignore start */
-import express from "express";
-import { body } from "express-validator";
-import { Logger } from "pino";
-
-import { MessageService } from "../services/message-service";
-import { NotificationService } from "../services/notification-service";
-import { errorMessage } from "../lib/errors";
-
-import type { AppContext } from "../index";
-
-import { initializeAgentFromSession } from "#/auth/session-agent";
-
-export class MessageController {
-  constructor(
-    private messageService: MessageService,
-    private logger: Logger,
-    private ctx: AppContext,
-    /**
-     * Optional push-notification side-effect. When wired (production routes),
-     * it triggers a push on new messages and drops subscriptions on account
-     * deletion. Optional + fire-and-forget so it never blocks a request, and
-     * existing call sites that omit it keep working unchanged.
-     */
-    private notificationService?: NotificationService
-  ) {}
-  /* v8 ignore stop */
-
-  /**
-   * Validation for adding example messages
-   */
-  validateAddExampleMessages = [
-    body("recipient").isString().notEmpty().withMessage("Recipient DID required"),
-  ];
-
-  /**
-   * Add example messages for a user
-   */
-  addExampleMessages = async (
-    req: express.Request,
-    res: express.Response
-  ): Promise<express.Response> => {
-    const recipient = req.session?.did;
-    if (!recipient) {
-      return res.status(403).json({ error: "Recipient DID required" });
-    }
-
-    try {
-      const messages = await this.messageService.addExampleMessages(recipient);
-      return res.json({ messages });
-    } catch (err) {
-      this.logger.error({ err, recipient }, "Failed to add example messages");
-      return res.status(500).json({ error: "Failed to add example messages" });
-    }
-  };
-
-  /**
-   * Validation for responding to a message
-   */
-  validateRespondToMessage = [
-    body("tid").isString().notEmpty().withMessage("Message TID required"),
-    body("recipient").isString().notEmpty().withMessage("Recipient DID required"),
-    body("original").isString().notEmpty().withMessage("Original message required"),
-    body("response")
-      .isString()
-      .isLength({ min: 1, max: 500 })
-      .withMessage("Response must be 1-500 chars"),
-    body("includeQuestionAsImage").isBoolean().optional(),
-    body("replyTo").optional().isObject(),
-    body("replyTo.uri").optional().isString(),
-    body("replyTo.cid").optional().isString(),
-  ];
-
-  /**
-   * Respond to a message and post to Bluesky
-   */
-  respondToMessage = async (
-    req: express.Request,
-    res: express.Response
-  ): Promise<express.Response> => {
-    const { tid, recipient, original, response, includeQuestionAsImage, replyTo } = req.body;
-
-    if (!tid || !recipient || !response) {
-      this.logger.warn({ tid, recipient, response }, "Missing required fields in respond endpoint");
-      return res.status(400).json({ error: "Missing required fields" });
-    }
-
-    const did = req.session?.did;
-    if (!did) {
-      this.logger.warn("No authenticated user session found");
-      return res.status(403).json({ error: "Not authenticated" });
-    }
-
-    const agent = await initializeAgentFromSession(req, this.ctx);
-    if (!agent) {
-      this.logger.warn({ did }, "No agent could be initialized from session");
-      return res.json({ isLoggedIn: false, profile: null, did: null });
-    }
-
-    try {
-      const result = await this.messageService.respondToMessage(
-        tid,
-        did,
-        recipient,
-        original,
-        response,
-        includeQuestionAsImage || false,
-        agent,
-        replyTo
-      );
-      return res.json(result);
-    } catch (err: unknown) {
-      this.logger.error(
-        { err, tid, did },
-        "Error in /messages/respond endpoint while trying to post to Bluesky"
-      );
-      return res.status(500).json({ error: errorMessage(err) || "Failed to post to Bluesky" });
-    }
-  };
-
-  /**
-   * Validation for sending a message
-   */
-  validateSendMessage = [
-    body("recipient").isString().notEmpty().withMessage("Recipient DID required"),
-    body("message")
-      .isString()
-      .isLength({ min: 1, max: 500 })
-      .withMessage("Message must be 1-500 chars"),
-  ];
-
-  /**
-   * Send an anonymous message
-   */
-  sendMessage = async (req: express.Request, res: express.Response): Promise<express.Response> => {
-    const { recipient, message } = req.body;
-
-    if (!recipient || !message) {
-      return res.status(400).json({ error: "Recipient and message required" });
-    }
-
-    try {
-      const result = await this.messageService.sendMessage(recipient, message);
-      // Per-send info log is on the hot path of every message send; demoted to
-      // debug so the info-level transport drops it (#319). The push side-effect
-      // below already logs failures at error level, which is what we need.
-      this.logger.debug({ recipient }, "Anonymous message sent");
-      // Fire-and-forget: a push failure must never affect the send response.
-      this.notificationService
-        ?.sendNewMessageNotification(recipient)
-        .catch((err) =>
-          this.logger.error({ err, did: recipient }, "Failed to send push notification")
-        );
-      return res.json(result);
-    } catch (err: unknown) {
-      this.logger.error({ err, recipient }, "Failed to send message");
-      const msg = errorMessage(err);
-      // "not found" → 404 (recipient isn't a Navyfragen user); "inbox is
-      // closed" → 403 (recipient exists but disabled intake). Everything else
-      // is an internal failure.
-      const status = msg.includes("not found") ? 404 : msg.includes("not accepting") ? 403 : 500;
-      return res.status(status).json({
-        error: msg || "Failed to send message",
-      });
-    }
-  };
-
-  /**
-   * Get messages for a user
-   */
-  getMessages = async (req: express.Request, res: express.Response): Promise<express.Response> => {
-    const recipient = req.session?.did;
-
-    if (!recipient) {
-      return res.status(403).json({ error: "Not authenticated" });
-    }
-
-    try {
-      const messages = await this.messageService.getMessages(recipient);
-      return res.json({ messages });
-    } catch (err: unknown) {
-      if (errorMessage(err).includes("not exist")) {
-        return res.status(404).json({ error: "User not found (user profile does not exist)" });
-      }
-      this.logger.error({ err, recipient }, "Failed to fetch messages");
-      return res.status(500).json({ error: "Failed to fetch messages" });
-    }
-  };
-
-  /**
-   * Delete a message
-   */
-  deleteMessage = async (
-    req: express.Request,
-    res: express.Response
-  ): Promise<express.Response> => {
-    const tid = req.params.tid as string;
-
-    if (!tid) {
-      return res.status(400).json({ error: "Message TID required" });
-    }
-
-    const userSessionDid = req.session?.did;
-    if (!userSessionDid) {
-      return res.status(403).json({ error: "Not authenticated" });
-    }
-
-    const agent = await initializeAgentFromSession(req, this.ctx);
-    if (!agent) {
-      this.logger.warn({ userSessionDid }, "No agent could be initialized from session");
-      return res.json({ isLoggedIn: false, profile: null, did: null });
-    }
-
-    try {
-      await this.messageService.deleteMessage(tid, userSessionDid, agent);
-      return res.json({ success: true });
-    } catch (err: unknown) {
-      const msg = errorMessage(err);
-      const status = msg.includes("not found") ? 404 : msg.includes("Not authorized") ? 403 : 500;
-
-      return res.status(status).json({ error: msg || "Failed to delete message" });
-    }
-  };
-
-  /**
-   * Delete all user data
-   */
-  deleteAccount = async (
-    req: express.Request,
-    res: express.Response
-  ): Promise<express.Response> => {
-    const userSessionDid = req.session?.did;
-
-    if (!userSessionDid) {
-      return res.status(403).json({ error: "Not authenticated" });
-    }
-
-    const agent = await initializeAgentFromSession(req, this.ctx);
-    if (!agent) {
-      return res.status(401).json({
-        error: "Authentication failed - could not initialize agent or retrieve user DID",
-      });
-    }
-
-    try {
-      await this.messageService.deleteUserData(userSessionDid, agent);
-      // Fire-and-forget: also drop any push subscriptions for this user.
-      this.notificationService
-        ?.deleteAllSubscriptionsForUser(userSessionDid)
-        .catch((err) =>
-          this.logger.error({ err, did: userSessionDid }, "Failed to delete push subscriptions")
-        );
-      req.session = null;
-      this.logger.info({ did: userSessionDid }, "Account and all data deleted");
-      return res.json({ success: true });
-    } catch (err: unknown) {
-      this.logger.error({ err, did: userSessionDid }, "Failed to delete account data");
-      return res.status(500).json({ error: errorMessage(err) || "Failed to delete account data" });
-    }
-  };
-
-  /**
-   * Sync messages to PDS
-   */
-  syncMessages = async (req: express.Request, res: express.Response): Promise<express.Response> => {
-    const userSessionDid = req.session?.did;
-
-    if (!userSessionDid) {
-      return res.status(403).json({ error: "Not authenticated" });
-    }
-
-    // Stopgap for now
-    const userSettings = await this.ctx.db
-      .selectFrom("user_settings")
-      .selectAll()
-      .where("did", "=", userSessionDid)
-      .executeTakeFirst();
-
-    if (!userSettings || !userSettings?.pdsSyncEnabled) {
-      return res.status(200).json({ success: true, message: "PDS sync is disabled" });
-    }
-
-    // Initialize the agent for the authenticated user session
-    const agent = await initializeAgentFromSession(req, this.ctx);
-
-    if (!agent) {
-      return res.status(401).json({
-        error: "Authentication failed - could not initialize agent",
-      });
-    }
-
-    try {
-      const syncResult = await this.messageService.syncMessages(userSessionDid, agent);
-      return res.json(syncResult);
-    } catch (err: unknown) {
-      this.logger.error({ err, did: userSessionDid }, "Failed to sync messages to PDS");
-      return res.status(500).json({
-        error: "Failed to sync messages to PDS",
-        details: errorMessage(err),
-      });
-    }
-  };
-  /* v8 ignore next 1 */
-}
